+++ tilematrix/__init__.py
@@ -2,7 +2,12 @@
 
 from ._tilepyramid import TilePyramid
 from ._tile import Tile
-from ._funcs import clip_geometry_to_srs_bounds, Shape, Bounds, TileIndex
+from ._funcs import clip_geometry_to_srs_bounds, snap_bounds, Shape, Bounds, TileIndex
+
+__all__ = [
+    'clip_geometry_to_srs_bounds', 'snap_bounds', 'Tile', 'TilePyramid', 'TileIndex',
+    'Shape', 'Bounds'
+]
 
 
-__version__ = "0.14.2"
+__version__ = "0.15"